Crate mithril_dmq

This crate provides mechanisms to publish and consume messages of a Decentralized Message Queue network through a DMQ node.

Modules

test

Test utilities.

Structs

DmqConsumerClientDeduplicator

A DMQ consumer client that filters out duplicate messages.

DmqConsumerClientPallas

A DMQ client consumer implementation.

DmqConsumerServerPallas

A DMQ server implementation for messages notification from a DMQ node.

DmqMessage

Wrapper for a DMQ message which can be serialized and deserialized.

DmqMessageBuilder

A builder for creating DMQ messages.

DmqPublisherClientPallas

A DMQ client publisher implementation.

DmqPublisherServerPallas

A DMQ server implementation for messages publication to a DMQ node.

SystemUnixTimestampProvider

Provides the current timestamp in seconds since the UNIX epoch from the system.

Enums

DmqNetwork

The Dmq Network that is being targeted

Constants

DMQ_MESSAGE_DEDUPLICATOR_TTL

Default maximum time to keep a seen message in the deduplicator cache.

Traits

DmqConsumerClient

Trait for the client side of consuming messages from a DMQ node.

DmqConsumerServer

Trait for the server side of consuming messages from a DMQ node.

DmqPublisherClient

Trait for the client side of publishing messages from a DMQ node.

DmqPublisherServer

Trait for the server side of publishing messages from a DMQ node.